ransomware-detector-khanh/requirements.txt
pefile>=2023.2.7
yara-python>=4.2.3
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
customtkinter>=5.2.0
reportlab>=4.0.0
joblib>=1.3.0
imbalanced-learn>=0.11.0
